pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "resilientops"
version = "0.1.0"
description = "Global supply chain crisis simulator for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic",
    "fastapi",
    "uvicorn"
]

[project.scripts]
server = "server.app:main"